2to3 first run

+++ land/mrfdev_pt1000.py
@@ -15,15 +15,15 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 '''
 
-from mrf_sens import MrfSens
-from mrf_dev  import MrfDev
-from mrf_sens_relay import MrfSensRelay
+from .mrf_sens import MrfSens
+from .mrf_dev  import MrfDev
+from .mrf_sens_relay import MrfSensRelay
 import datetime
 import ctypes
-from mrf_structs import *
-from core_tests import mrf_cmd_app_test
+from .mrf_structs import *
+from .core_tests import mrf_cmd_app_test
 from math import sqrt
-from mrflog import mrflog
+from .mrflog import mrflog
 from collections import deque
 
 
@@ -150,7 +150,7 @@ Pt1000AppCmds = {
 
 class MrfSensPt1000(MrfSens):
     _in_flds_ = [ ('date', PktTimeDate) ,
-                  ('milliohms' , long) ]  # hmpff
+                  ('milliohms' , int) ]  # hmpff
 
     _out_flds_ = [ ('send_date' , datetime.datetime.now ),
                    ('recd_date' , datetime.datetime.now),
